@@ -5,39 +5,34 @@
 	"repository": "https://github.com/project-trio/td",
 	"description": "Top-down 3D, multiplayer tower defense browser game.",
 	"scripts": {
-		"i": "pnpm install",
 		"serve": "vue-cli-service serve",
 		"build": "vue-cli-service build --modern"
 	},
-	"engines": {
-		"node": "^12",
-		"npm": "^6"
-	},
 	"devDependencies": {
-		"@fullhuman/postcss-purgecss": "^1.3.0",
-		"@ky-is/eslint-config": "^1.7.0",
+		"@fullhuman/postcss-purgecss": "^2.3.0",
+		"@ky-is/eslint-config": "^2.1.0",
 		"@ky-is/tailwind-color-palette": "^1.0.0",
 		"@ky-is/tailwindcss-plugin-width-height": "^1.0.0",
 		"@ky-is/trio-client": "^1.0.0",
 		"@ky-is/vue-cli-plugin-tabs": "^1.1.0",
 		"@ky-is/vue-cli-plugin-tailwind": "^2.0.0",
-		"@vue/babel-preset-app": "^4.1.2",
-		"@vue/cli-plugin-babel": "^4.1.2",
-		"@vue/cli-plugin-eslint": "^4.1.2",
-		"@vue/cli-service": "^4.1.2",
-		"babel-eslint": "^10.0.3",
-		"core-js": "^3.6.1",
-		"eslint": "^6.8.0",
-		"eslint-plugin-vue": "^6.1.1",
+		"@vue/babel-preset-app": "^4.4.6",
+		"@vue/cli-plugin-babel": "^4.4.6",
+		"@vue/cli-plugin-eslint": "^4.4.6",
+		"@vue/cli-service": "^4.4.6",
+		"babel-eslint": "^10.1.0",
+		"core-js": "^3.6.5",
+		"eslint": "^7.6.0",
+		"eslint-plugin-vue": "^7.0.0-beta.1",
 		"postcss-preset-env": "^6.7.0",
 		"seedrandom": "^3.0.5",
 		"stats.js": "^0.17.0",
-		"tailwindcss": "^1.1.4",
-		"three": "^0.112.1",
-		"typescript": "^3.7.4",
-		"url-loader": "^3.0.0",
+		"tailwindcss": "^1.6.2",
+		"three": "^0.119.1",
+		"typescript": "^3.9.7",
+		"url-loader": "^4.1.0",
 		"vue": "^2.6.11",
-		"vue-router": "^3.1.3",
+		"vue-router": "^3.3.4",
 		"vue-template-compiler": "^2.6.11"
 	},
 	"babel": {
@@ -65,9 +60,10 @@
 	},
 	"eslintConfig": {
 		"extends": "@ky-is/eslint-config/vue",
-		"parserOptions": {
-			"parser": "babel-eslint"
-		},
-		"rules": {}
+		"rules": {
+			"vue/html-indent": ["warn", "tab", {
+				"baseIndent": 0
+			}]
+		}
 	}
 }
