@@ -134,7 +134,7 @@ export default {
 		window.addEventListener('keydown', this.keydown)
 	},
 
-	beforeDestroy () {
+	beforeUnmount () {
 		window.removeEventListener('keydown', this.keydown)
 	},
 
