@@ -1,7 +1,7 @@
 <template>
 <Popover class="wave-box  relative pl-2 pb-px bg-contain bg-center bg-no-repeat capitalize  flex items-end" :style="{ 'background-image': `url(${require(`@/app/assets/icons/${name}.png`)})` }">
 	{{ number }}
-	<template slot="hover">{{ name }} {{ boss ? 'Boss' : null }}</template>
+	<template #hover>{{ name }} {{ boss ? 'Boss' : null }}</template>
 </Popover>
 </template>
 
